api/requirements.txt
blinker==1.4
click==6.6
Flask==0.11.1
Flask-Cors==3.0.2
Flask-SocketIO==2.7.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-engineio==1.0.3
python-socketio==1.6.0
RPi.GPIO==0.6.2
six==1.10.0
Werkzeug==0.11.11
